Extract WikiArticleSearchTask from WikiArticleHelper
